Tentative to fix #11 and #16 using a hashcode.
For markers, it's something like a «Minimal perfect hashing» (just an incremented counter)
For clusters, collisions are possible (Birthday paradoxe) but while the hashing function is far from perfect (yet) the probability of a collision is very low, and not critical. The check on the population helps too.

WIP : not well tested and LeafletAdapter.ts needs a refactoring