Don't use insertionSort when more than 300 markers are added
Uses the old 20% rule up to 300 and then always does the built-in (n log n) sort.

This is a big performance win for large datasets when eg. 19000 new markers could trigger an
insertionSort on a 100k dataset.

Measured on a real application: sorting time during app startup dropped from 70seconds(!!) to <200ms.