Split build grunt task into dev and dist targets
The main difference is that the dev target generates a source map while
the build target does not. Another difference is that the dist code is
minified while the dev code is not.

The build dir is added to .gitignore since this is only used for
development and should not be distributed.